refactored non-centroid-pois in multple files
